Extracted currentUser state from App to UserContext.js, changed the name of context
