+++ client/src/Pages/Home/Home.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import BusinessList from '../../../src/components/BusinessList/BusinessList';
 import SearchBar from '../../../src/components/SearchBar/SearchBar';
-import Yelp from '../../util/yelp';
+import Yelp from '../../util/Api/yelp';
 
 const Home = () => {
     const [businesses, setBusinesses] = useState([]);
